Fixed Healthbars and ammocount. Also added healthbars on top of zombies.

+++ Assets/Scripts/Character_Controller.cs
@@ -12,7 +12,7 @@ public class Character_Controller : MonoBehaviour
     public bool magazineEmpty = false;
 
     //variables for raycast
-    public float damage = 10f;
+    public int damage = 10;
     public float range = 100f;
 
     public Camera fpsCam;
